// src/app/feature-modules/home/types/home.contracts.types.ts
/*
export interface Contract {
  pk: string;
  item_seg_id: string;
  contract_id: string;
  sk: string;
  district: string;
  project_type: string;
  status: string;
  project_desc: string;
  total_amount: string;
  contract_days: string;
  work_begin_date: string;
  final_acceptance_date: string;
  est_work_begin_date: string;
  est_final_acceptance_date: string;
  act_contract_days: string;
  est_contract_days: string;
  letting_date: string;
}

// export interface Contract {
//   contract_id: string;
//   sortKey: string;
//   district: number;
//   project_type: string;
//   status: string;
//   description: string;
//   center_Lane_Miles: number;
//   lane_Miles: number;
//   original_Contract_Amount: number;
//   total_amount: number;
//   contract_days: number;
//   days_Used: number;
//   award_Date: Date;
//   work_begin_date: Date;
//   final_acceptance_date: Date;
//   primary_County: string;
// }

export interface Contracts {
  attributes: Contract[];
  fields: {
    name: string;
    alias: string;
    type?: string;
  }[];
  hasNoFilterResults: boolean;
}
*/